fix: Add typecheck scripts and fix TypeScript errors
- Add typecheck scripts to backend, frontend, and root package.json
- Fix customer.routes.ts type errors (userId string|bigint -> bigint cast)
- Delete unused App.tsx test file from Iteration 1

All TypeScript checks now pass.

+++ coab-backend/src/routes/customer.routes.ts
@@ -39,7 +39,7 @@ const customerRoutes: FastifyPluginAsync = async (fastify) => {
      */
     protectedRoutes.get('/me', async (request, reply) => {
       try {
-        const clienteId = request.user!.userId;
+        const clienteId = request.user!.userId as bigint;
         const profile = await customerService.getCustomerProfile(clienteId);
         return profile;
       } catch (error: any) {
@@ -61,7 +61,7 @@ const customerRoutes: FastifyPluginAsync = async (fastify) => {
      */
     protectedRoutes.get('/me/saldo', async (request, reply) => {
       try {
-        const clienteId = request.user!.userId;
+        const clienteId = request.user!.userId as bigint;
         const balance = await customerService.getCustomerBalance(clienteId);
         return balance;
       } catch (error) {
@@ -79,7 +79,7 @@ const customerRoutes: FastifyPluginAsync = async (fastify) => {
     protectedRoutes.get('/me/pagos', async (request, reply) => {
       try {
         const query = paginationSchema.parse(request.query);
-        const clienteId = request.user!.userId;
+        const clienteId = request.user!.userId as bigint;
 
         const result = await customerService.getCustomerPayments(
           clienteId,
@@ -111,7 +111,7 @@ const customerRoutes: FastifyPluginAsync = async (fastify) => {
     protectedRoutes.get('/me/boletas', async (request, reply) => {
       try {
         const query = paginationSchema.parse(request.query);
-        const clienteId = request.user!.userId;
+        const clienteId = request.user!.userId as bigint;
 
         const result = await customerService.getCustomerBoletas(
           clienteId,
@@ -143,7 +143,7 @@ const customerRoutes: FastifyPluginAsync = async (fastify) => {
     protectedRoutes.get('/me/boletas/:id', async (request, reply) => {
       try {
         const params = boletaIdSchema.parse(request.params);
-        const clienteId = request.user!.userId;
+        const clienteId = request.user!.userId as bigint;
 
         const boleta = await customerService.getBoletaById(
           clienteId,
